+++ docs/common/conf.py
@@ -59,7 +59,6 @@ nitpick_ignore = [
     ('py:class', 'poptorch.MatMulSerializationMode'),
     ('py:class', 'poptorch.ReductionType'),
     ('py:class', 'poptorch.SyncPattern'),
-    ('py:class', 'poptorch.DataLoaderMode'),
 ]
 
 # -- Options for HTML output -------------------------------------------------