+++ docs/user_guide/changelog.rst
@@ -10,6 +10,7 @@ New features
 
 - Add support for ``torch.unbind``
 - Add option to set `poptorch.Options` using options specified in a config file.
+- Added ``mode=poptorch.DataLoaderMode.AsyncRebatched``
 
 v2.0 (Poplar SDK 2.0)
 =====================